// src/interfaces/cards/UnnormalizedCard.ts
export interface UnnormalizedCard {
    title: string;
    subtitle: string;
    description: string;
    phone: string;
    email: string;
    wed?: string;
    url?: string;
    alt?: string;
    state?: string;
    country: string;
    city: string;
    street: string;
    houseNumber: number;
    zip: number;
    
}